Update outp_pkgs.rst
Added row in the diagnostics "codes" table for 'W' points. There were codes for 'S', 'U', and 'V' but not 'W'.  The problem is that many diagnostics use the 'W' code and the meaning of 'W' was not defined. Also added another example for the diag code: "WM LR", for 'w' point, 'center', 'top', and 'Nr' vertical levels.